Revert "ModuleComparator"
This reverts commit 11221af8b9153c99e0d93086b326d1efaf8990eb.
